pyproject.toml
[project]
name = "guardials"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "openai-agents>=0.2.4",
    "rich>=14.1.0",
]
